LangGraph/requirements.txt
# LangGraph module dependencies
# If using the root venv, these are already installed.
# Only needed if running this module standalone.
langchain>=0.3.0
langchain-ollama>=0.1.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
python-dotenv>=1.0.0
